Update homepage.html
I added photos to indicate the contents of the lessons page and the stories page. 

Also included the addition of the flask if statement to check whether a user is logged in. If they are, the logout button will be displayed, if not then the login and sign up buttons will be shown